chore: move mock data to root component
