+++ src/components/main/styles.ts
@@ -5,6 +5,7 @@ const padding = 20;
 export const styles = StyleSheet.create({
   root: {
     backgroundColor: 'white',
+    height: '100%',
   },
   header: {
     padding: 16,
